requirements_py3.7+.txt
beautifulsoup4==4.9.0
bs4==0.0.1
Flask==0.12.2
gensim==3.8.2
google==2.0.3
matplotlib==3.2.1
networkx==2.1
nltk==3.4.5 # nltk.download('stopwords/wordnet/punkt')
numpy==1.17.4
pandas==0.25.3
pyspotlight==0.7.2
requests==2.18.4
scipy==1.4.1
spacy==2.2.3
SPARQLWrapper==1.8.5
stanford-openie==1.0.1
stanfordnlp==0.2.0
torch==1.5.0+cpu # pip install torch==1.6.0+cpu torchvision==0.7.0+cpu -f https://download.pytorch.org/whl/torch_stable.html
tqdm==4.45.0
urllib3==1.22
Wikipedia-API==0.5.4
# java installed and in environment path
# wikifer key in api_async